query builder: regoranize helpers
