+++ .changeset/pink-pots-pay.md
@@ -1,0 +1,8 @@
+---
+'@finos/legend-application': patch
+'@finos/legend-application-query': patch
+'@finos/legend-application-query-bootstrap': patch
+'@finos/legend-application-studio': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-server-sdlc': patch
+---
