+++ packages/legend-application-query/src/stores/fetch-structure/projection/post-filter/operators/QueryBuilderPostFilterOperator_Contain.ts
@@ -27,18 +27,18 @@ import {
   UnsupportedOperationError,
 } from '@finos/legend-shared';
 import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../../../../QueryBuilder_Const.js';
-import {
-  buildNotExpression,
-  getNonCollectionValueSpecificationType,
-  unwrapNotExpression,
-} from '../../../../QueryBuilderOperatorsHelper.js';
 import { QueryBuilderPostFilterOperator } from '../QueryBuilderPostFilterOperator.js';
 import { buildPostFilterConditionState } from '../QueryBuilderPostFilterValueSpecificationProcessor.js';
 import type {
   PostFilterConditionState,
   QueryBuilderPostFilterState,
 } from '../QueryBuilderPostFilterState.js';
-import { generateDefaultValueForPrimitiveType } from '../../../../QueryBuilderValueSpecificationBuilderHelper.js';
+import {
+  buildNotExpression,
+  generateDefaultValueForPrimitiveType,
+  getNonCollectionValueSpecificationType,
+  unwrapNotExpression,
+} from '../../../../QueryBuilderValueSpecificationHelper.js';
 import { buildPostFilterConditionExpression } from './QueryBuilderPostFilterOperatorHelper.js';
 
 export class QueryBuilderPostFilterOperator_Contain extends QueryBuilderPostFilterOperator {
