+++ packages/legend-application-query/src/stores/filter/operators/QueryBuilderFilterOperatorHelper.ts
@@ -35,7 +35,7 @@ import {
   type QueryBuilderFilterState,
 } from '../QueryBuilderFilterState.js';
 import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../../QueryBuilder_Const.js';
-import { buildGenericLambdaFunctionInstanceValue } from '../../QueryBuilderValueSpecificationBuilderHelper.js';
+import { buildGenericLambdaFunctionInstanceValue } from '../../QueryBuilderValueSpecificationHelper.js';
 import { buildPropertyExpressionChain } from '../../QueryBuilderValueSpecificationBuilder.js';
 import type { QueryBuilderFilterOperator } from '../QueryBuilderFilterOperator.js';
